refactor(popup): move ConnectWallet route out of Settings
